add basic network and data interaction

+++ app/src/main/java/ru/nikshlykov/rickandmortyapiapp/ui/model/Character.kt
@@ -1,10 +1,10 @@
 package ru.nikshlykov.rickandmortyapiapp.ui.model
 
 data class Character(
-  val id: Long,
+  val id: Int,
   val name: String,
   val imageUrl: String,
   val additionalData: String
 )
 
-// TODO подумать над тем, как сделать тут Bitmap, а не ссылку.
+// TODO подумать над тем, как сделать тут Bitmap/Drawable, а не ссылку.
